- Desenvolvimento das camadas de aplicação, apresentação, e core - Separação de camadas entre diferentes projetos de uma mesma solução - Adicionadas classes já existentes em suas respectivas camadas.
