+++ TheatricalPlayersRefactoringKata.Application/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace TheatricalPlayersRefactoringKata.Application;
+
+public class Class1
+{
+
+}
