+++ TheatricalPlayersRefactoringKata.Core/Entities/Performance.cs
@@ -1,4 +1,4 @@
-namespace TheatricalPlayersRefactoringKata;
+namespace TheatricalPlayersRefactoringKata.Core.Entities;
 
 public class Performance
 {
