+++ TheatricalPlayersRefactoringKata.Core/Entities/Play.cs
@@ -1,4 +1,4 @@
-namespace TheatricalPlayersRefactoringKata;
+namespace TheatricalPlayersRefactoringKata.Core.Entities;
 
 public class Play
 {
@@ -10,7 +10,8 @@ public class Play
     public int Lines { get => _lines; set => _lines = value; }
     public string Type { get => _type; set => _type = value; }
 
-    public Play(string name, int lines, string type) {
+    public Play(string name, int lines, string type)
+    {
         this._name = name;
         this._lines = lines;
         this._type = type;
